+++ Pod/Classes/ViewControllers/ARCityBMWListComponentViewController.m
@@ -1,0 +1,10 @@
+#import "ARCityBMWListComponentViewController.h"
+
+@implementation ARCityBMWListComponentViewController
+
+- (instancetype)initWithCitySlug:(NSString *)citySlug
+{
+    return [super initWithEmission:nil moduleName:@"CitySectionList"];
+}
+
+@end